Fixed the NECTAR routing protocol
  * The protocol specifies that the *meeting frequency* be determined, not the
    count. Therefore the data in the routing table had to change, so that the
    frequency is calculated from the count and the time of the first encounter
    with a particular node.
  * I also fixed the other code that supports the NECTAR protocol to fit with
    the new (fixed) implementation.

+++ app/src/main/java/org/radindustries/radwolfsdragon/examples/wifip2ppeerdiscoverytest/dtn/daemon/NECTARRouter2Daemon.java
@@ -3,5 +3,5 @@ package org.radindustries.radwolfsdragon.examples.wifip2ppeerdiscoverytest.dtn.d
 import org.radindustries.radwolfsdragon.examples.wifip2ppeerdiscoverytest.dtn.dto.DTNEndpointID;
 
 public interface NECTARRouter2Daemon extends Router2Daemon {
-    int getMeetingFrequency(DTNEndpointID nodeEID);
+    float getMeetingFrequency(DTNEndpointID nodeEID);
 }
