+++ app/src/main/java/org/radindustries/radwolfsdragon/examples/wifip2ppeerdiscoverytest/dtn/router/Daemon2NECTARRoutingTable.java
@@ -3,6 +3,6 @@ package org.radindustries.radwolfsdragon.examples.wifip2ppeerdiscoverytest.dtn.r
 import org.radindustries.radwolfsdragon.examples.wifip2ppeerdiscoverytest.dtn.dto.DTNEndpointID;
 
 public interface Daemon2NECTARRoutingTable {
-    int getMeetingFrequency(DTNEndpointID nodeEID);
+    float getMeetingFrequency(DTNEndpointID nodeEID);
     void incrementMeetingFrequency(DTNEndpointID nodeEID);
 }
